Orca code for Implementation of Booth's Algorithm
This code is not robust, but a map for binary addition and Arithmetic Shift right (ASR). Credits to Qijun (Mary) Hu and Professor David Perkins for the binary addition code which was completed in a mini-project where we tried to implement things in Orca that it is not designed for.